pyproject.toml
[project]
name = "awesome-nano-banana-search"
version = "0.1.0"
description = "Search engine for awesome-nano-banana-index"
readme = "search/README.md"
requires-python = ">=3.8"
dependencies = [
    "fastapi==0.104.1",
    "uvicorn[standard]==0.24.0",
    "meilisearch==0.32.0",
    "pyyaml==6.0.1",
    "click==8.1.7",
    "python-dotenv==1.0.0",
    "jinja2==3.1.2",
    "aiofiles==23.2.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = []

[tool.black]
line-length = 100
target-version = ['py38']

[tool.ruff]
line-length = 100
target-version = "py38"
